Full mobile page support
I (painstakingly) remade all the pages so that they work on mobile devices. They will likely need to be redone in the future to be more aesthetically pleasing, but as of right now they at least work on mobile devices. I also redesigned the Sign In/Sign Up page and fixed a title renaming bug (that I could've sworn I already fixed before).