Implemented local storage for user history
I realized that our read count was ballooning because of how fetching user history was implemented. I changed it so that user data is fetched once and then put into local storage so that history can be viewed and edited without reading from firebase every single time. This should keep our read count much lower (which is necessary if we want to grow our user base).